clean up test files and shared code
